Refactored src and src/components directories
- created MainPage directory in src/components and moved relevant files into it

- moved src/app/store.js to src/store.js (unnecessary nesting)
