+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { createMuiTheme, MuiThemeProvider } from '@material-ui/core'
-import './background.css'
+import './components/MainPage/background.css'
 import { Provider } from 'react-redux'
-import store from './app/store'
+import store from './store'
 import { Auth0Provider } from './components/react-auth0-spa'
 import history from './utils/history'
 import { useDarkMode } from './customHooks/useDarkMode'
