+++ src/components/AppBarComponent.js
@@ -3,7 +3,7 @@ import history from '../utils/history'
 import LeftMenuBar from './LeftMenuBar'
 import Grid from '@material-ui/core/Grid'
 import React from 'react'
-import MainPage from './MainPage'
+import MainPage from './MainPage/MainPage'
 import { HomePage } from './HomePage/HomePage'
 import { EraseQuestionnaireButton } from '../navigationBarButtons/EraseQuestionnaireButton'
 import { RenderQuestionnaireButton } from '../navigationBarButtons/RenderQuestionnaireButton'
