+++ src/components/MainPage/EditDialogTitle.js
@@ -1,11 +1,11 @@
 import React, {useEffect} from "react";
 import {useDispatch} from "react-redux";
-import {UPDATE_QUESTION} from "../features/questions/questionsSlice";
-import store from "../app/store";
-import {CLEAN_SUPER_QUESTION} from "../utils";
-import {postprocessQuestion} from "./properties/postprocessor";
-import {TitleProperty} from "./properties/TextProperties";
-import {setQuestion} from "../features/question/questionSlice";
+import {UPDATE_QUESTION} from "../../features/questions/questionsSlice";
+import store from "../../store";
+import {CLEAN_SUPER_QUESTION} from "../../utils";
+import {postprocessQuestion} from "../properties/postprocessor";
+import {TitleProperty} from "../properties/TextProperties";
+import {setQuestion} from "../../features/question/questionSlice";
 
 /**
  * In general, this is a stripped down version of the edit dialog, containing one property: title.
