+++ src/components/MainPage/EditingFeature.js
@@ -1,14 +1,14 @@
 import React, {createElement, useEffect} from "react";
 import {Box, Button, Dialog, Divider, Grid, Paper, Typography,} from "@material-ui/core";
-import {CLEAN_SUPER_QUESTION, PROPERTIES_BY_QUESTION_TYPE,} from "../utils";
+import {CLEAN_SUPER_QUESTION, PROPERTIES_BY_QUESTION_TYPE,} from "../../utils";
 import {useDispatch, useSelector} from "react-redux";
-import {setQuestion,} from "../features/question/questionSlice";
-import TypeProperty from "./properties/TypeProperty";
-import store from "../app/store";
-import {postprocessQuestion} from "./properties/postprocessor";
-import {UPDATE_QUESTION} from "../features/questions/questionsSlice";
-import {SET_UTILITIES} from "../features/utilities/utilitiesSlice";
-import {useStyles} from "./SaveQuestionnaireDialog"
+import {setQuestion,} from "../../features/question/questionSlice";
+import TypeProperty from "../properties/TypeProperty";
+import store from "../../store";
+import {postprocessQuestion} from "../properties/postprocessor";
+import {UPDATE_QUESTION} from "../../features/questions/questionsSlice";
+import {SET_UTILITIES} from "../../features/utilities/utilitiesSlice";
+import {useStyles} from "../SaveQuestionnaireDialog"
 
 
 const EditDialog2 = ({question, index, open, setOpen}) => {
