+++ src/components/MainPage/ExpansionRule.js
