+++ src/components/MainPage/MainPage.js
@@ -1,15 +1,15 @@
 import React, {useEffect} from 'react'
 import {CssBaseline, Grid, makeStyles} from "@material-ui/core";
 import {DragDropContext} from "react-beautiful-dnd";
-import "./index.css";
+import "../index.css";
 import {QUESTION_TYPES} from "./QuestionTypes";
-import {QuestionTypesMenu} from "../QuestionTypesMenu";
-import {QuestionsArea} from "../QuestionsArea";
-import {JSONTranslationArea} from "../JSONTranslationArea";
+import {QuestionTypesMenu} from "./QuestionTypesMenu";
+import {QuestionsArea} from "./QuestionsArea";
+import {JSONTranslationArea} from "./JSONTranslationArea";
 import {useDispatch, useSelector} from "react-redux";
-import { CLEAR_MAPS, initializeMaps } from '../features/utilities/utilitiesSlice'
-import BackToTopArrowButton from "../BackToTopArrowButton";
-import useDrag from '../customHooks/useDrag'
+import { CLEAR_MAPS, initializeMaps } from '../../features/utilities/utilitiesSlice'
+import BackToTopArrowButton from "./BackToTopArrowButton";
+import useDrag from '../../customHooks/useDrag'
 
 const useStyles = makeStyles(theme => ({
     root: {
