+++ src/components/MainPage/QuestionTypes.js
