+++ src/components/MainPage/QuestionTypesMenu.js
