+++ src/components/MainPage/background.css
