+++ src/components/buttons/EditQuestionButton.js
@@ -1,8 +1,8 @@
 import React, {useState} from "react";
 import {IconButton, Tooltip} from "@material-ui/core";
 import EditIcon from "@material-ui/icons/Edit";
-import EditDialog2 from "../EditingFeature";
-import store from "../../app/store";
+import EditDialog2 from "../MainPage/EditingFeature";
+import store from "../../store";
 import {SET_SAVED} from "../../features/utilities/utilitiesSlice";
